TestCapacityScheduler: apps management tests moved to separate file
